// src/types/iconTypes.ts
// src/types/iconTypes.ts

export type IconName =
  | 'user'
  | 'phone'
  | 'email'
  | 'plus'
  | 'close'
  | 'facebook'
  | 'instagram'
  | 'x'
  | 'tikTok' 
  | 'linkedIn' 
  | 'website'
  | 'viber'
  | 'whatsApp' 
  | 'telegram'
  | 'signal' 
  | 'edit'
  | 'trash'
  | 'search'
  | 'dashboard'
  | 'usersGroup'
  | 'settings'
  | 'chevronDoubleLeft'
  | 'chevronDoubleRight' 
  | 'spinner'
  | 'alertTriangle'
  | 'info' 
  | 'successCircle' 
  | 'errorCircle' 
  | 'eye' 
  | 'eyeSlash' 
  | 'minus' 
  | 'pin' 
  | 'pinOutline' 
  | 'paperclip'
  | 'fileText'
  | 'fileImage'
  | 'filePdf'
  | 'fileArchive'
  | 'fileGeneric'
  | 'sparkles' // <<< ADDED
  | 'bell'
  | 'checkCircle' 
  | 'filter'
  | 'formatBold'
  | 'formatItalic'
  | 'listBullet'
  | 'queueList'
  | 'phoneHangup'    
  | 'phoneAnswer'    // Placeholder if needed, currently not used directly
  | 'microphone'     
  | 'microphoneSlash'
  | 'pauseCircle'    
  | 'playCircle'     // Placeholder if Hold is a toggle, currently using PauseCircle
  | 'dialpad'
  | 'cog' 
  | 'clipboardDocument' 
  | 'paperAirplane'     
  | 'arrowUturnLeft' // For backspace/delete in dialpad
  | 'chartBar' 
  ;